Tooling: Building Caching improvements

+++ Scripts/init/stage1.py
@@ -40,45 +40,48 @@ match OS:
         path_delimiter = "/"
         rar = ".tar.gz"
 
-tools_dir = sys.argv[2]
+tools_dir = os.getcwd() + path_delimiter + sys.argv[2]
+deps_dir = sys.argv[3]
 
 # Tools urls
-choosenim_dir = os.getcwd() + path_delimiter + tools_dir + path_delimiter + "nim"
+choosenim_dir = tools_dir + path_delimiter + "nim"
 
 # temporary
 choosenim_exe = "choosenim-" + choosenim_version + "_" + platform + "_amd64" + exe
+choosenim_cache = deps_dir + path_delimiter + choosenim_exe
 choosenim_url = "https://github.com/nim-lang/choosenim/releases/download/v" + choosenim_version + "/" + choosenim_exe
 choosenim_command = choosenim_dir + path_delimiter + choosenim_exe
 
-def download_choosenim(choosenim_url, choosenim_version):
-    print("Downloading Nim Bootstrapper")
-    global cwd
-    kill_file = os.getcwd() + path_delimiter + "Tools" + path_delimiter + "stage1"
-    if OS == "win":
-        query_parameters = {"downloadFormat": "exe"}
-    else:
-        query_parameters = {"downloadFormat": "bin"}
-    response = requests.get(choosenim_url, params=query_parameters)
-    if response.ok and response.status_code == 200:
-        with open(choosenim_command, mode="wb") as file:
-            file.write(response.content)
-    else:
-        print("Nim files failed to download; please check your internet connection and try again later")
-        with open(kill_file, mode="w") as file:
-            file.write("")
+def download_choosenim(choosenim_url):
+    if not pathlib.Path(choosenim_cache).exists():
+        print("Downloading Nim Bootstrapper")
+        global cwd
+        kill_file = os.getcwd() + path_delimiter + "Tools" + path_delimiter + "stage1"
+        if OS == "win":
+            query_parameters = {"downloadFormat": "exe"}
+        else:
+            query_parameters = {"downloadFormat": "bin"}
+        response = requests.get(choosenim_url, params=query_parameters)
+        if response.ok and response.status_code == 200:
+            with open(choosenim_command, mode="wb") as file:
+                file.write(response.content)
+        else:
+            print("Nim files failed to download; please check your internet connection and try again later")
+            with open(kill_file, mode="w") as file:
+                file.write("")
 
 # Create Nim directory if doesn't exist
-choosenim_dir = pathlib.Path(choosenim_dir)
+choosenim_dir = pathlib.Path(choosenim_dir).__str__()
 if not pathlib.Path(choosenim_dir).exists():
     pathlib.Path(choosenim_dir).mkdir(parents=True, exist_ok=True)
-    download_choosenim(choosenim_url, choosenim_version)
+    download_choosenim(choosenim_url)
 
 # For convenience and sanity
 choosenim_dir = pathlib.Path(choosenim_dir).__str__()
 if pathlib.Path(choosenim_command).exists():
-    pathlib.Path(choosenim_command).replace(choosenim_dir + path_delimiter + "choosenim" + exe)
+    pathlib.Path(choosenim_command).replace(choosenim_cache)
 
-choosenim_command = choosenim_dir + path_delimiter + "choosenim" + exe
+choosenim_command = choosenim_cache
 
 # Download nim versions
 def download_nim(nim_version):
